requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
chromadb>=0.5.0
httpx>=0.27.0
python-multipart>=0.0.18
pymupdf>=1.24.0
python-docx>=1.1.0
ebooklib>=0.18
beautifulsoup4>=4.12.0
sse-starlette>=2.1.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
fpdf2>=2.7.0
